refactor(decompose): split webapp/e2e-tests/playwright-browser/for_feature_development_skip/graph-rendering/node-placement-overlap-avoidance.spec.ts into folder
