+++ webapp/e2e-tests/playwright-browser/for_feature_development_skip/graph-rendering/node-placement-overlap-avoidance/testFixture.ts
@@ -1,0 +1,44 @@
+import { test as base } from '@playwright/test';
+
+type ConsoleCapture = {
+  consoleLogs: string[];
+  pageErrors: string[];
+  testLogs: string[];
+};
+
+// Console capture fixture (same pattern as other tests)
+export const test = base.extend<{ consoleCapture: ConsoleCapture }>({
+  consoleCapture: async ({ page }, use, testInfo) => {
+    const consoleLogs: string[] = [];
+    const pageErrors: string[] = [];
+    const testLogs: string[] = [];
+
+    page.on('console', msg => {
+      consoleLogs.push(`[Browser ${msg.type()}] ${msg.text()}`);
+    });
+
+    page.on('pageerror', error => {
+      pageErrors.push(`[Browser Error] ${error.message}\n${error.stack ?? ''}`);
+    });
+
+    const originalLog = console.log;
+    console.log = (...args: unknown[]) => {
+      testLogs.push(args.map(arg => String(arg)).join(' '));
+    };
+
+    await use({ consoleLogs, pageErrors, testLogs });
+
+    console.log = originalLog;
+
+    if (testInfo.status !== 'passed') {
+      console.log('\n=== Test Logs ===');
+      testLogs.forEach(log => console.log(log));
+      console.log('\n=== Browser Console Logs ===');
+      consoleLogs.forEach(log => console.log(log));
+      if (pageErrors.length > 0) {
+        console.log('\n=== Browser Errors ===');
+        pageErrors.forEach(err => console.log(err));
+      }
+    }
+  }
+});
